requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
python-dotenv==1.0.1
scikit-learn==1.6.0
scipy==1.14.1
threadpoolctl==3.5.0
Werkzeug==3.1.3
